+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # imv Change Log
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented here.
+
+## 1.3.0
+
+- **Added:** `--ignore` option to ignore files that match this glob pattern.
 
 ## 1.2.0